unify the selector callback and move plot functions to a new file
